requirements.txt
alembic==1.10.2
aniso8601==9.0.1
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.5
pure-eval==0.2.2
Pygments==2.15.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
pywin32==306
pyzmq==25.0.2
six==1.16.0
speedtest-cli==2.1.3
SQLAlchemy==2.0.4
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
Werkzeug==2.2.3
